Proyecto_AG_C2/requirements.txt
# FastAPI Framework y servidor ASGI
fastapi==0.109.2
uvicorn[standard]==0.27.0

# MongoDB - Driver asíncrono
motor==3.3.2
pymongo==4.6.1

# Validación de datos
pydantic==2.6.1
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0

# Logging mejorado
python-json-logger==2.0.7

# HTTP client (para health checks en Docker)
requests==2.31.0
